% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_ensemble_verification.R
\name{aggr_lead_plus_3months}
\alias{aggr_lead_plus_3months}
\title{aggregate forecast/observation for January+lead by polygon}
\usage{
aggr_lead_plus_3months(x, pol, lead)
}
\description{
aggregate forecast/observation for January+lead by polygon
}
