% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_ensemble_verification.R
\name{aggr_month_after_lead}
\alias{aggr_month_after_lead}
\title{aggregate forecast/observation for January+lead by polygon}
\usage{
aggr_month_after_lead(x, pol, lead)
}
\description{
aggregate forecast/observation for January+lead by polygon
}
