% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_ensemble_verification.R
\name{apply_roc}
\alias{apply_roc}
\title{function apply ROC}
\usage{
apply_roc(pol, xfc, xobs, model, lead, plower, pupper)
}
\arguments{
\item{pol}{is the polygon for which xfc and xobs will be aggregated}

\item{xfc}{is a spacetime data-frame with the forecast}

\item{xobs}{is a scapetime data-frame with the observation}

\item{model}{is the name of the model, eg "echam+xds" or "echam+eqm" or "mmsf"}

\item{lead}{is an integer with the value of the lead time of the prediction in months}
}
\description{
function apply ROC
}
