% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detach_package.R
\name{detach_package}
\alias{detach_package}
\title{detach package}
\usage{
detach_package(pkg)
}
\arguments{
\item{pkg}{the package name as string}
}
\description{
This function allows you to detach/unload a package that has been loaded.
}
\examples{
detach_package("roxygen2")
}
\keyword{detach}
\keyword{package}
\keyword{unload}
