% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_ensemble_verification.R
\name{get_cond_av_obs}
\alias{get_cond_av_obs}
\title{conditional average observation as in wilks page 286 eq 7.38}
\usage{
get_cond_av_obs(xfc_bins, event_tbl)
}
\description{
conditional average observation as in wilks page 286 eq 7.38
}
