% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_ensemble_verification.R
\name{get_fc_prob_bins}
\alias{get_fc_prob_bins}
\title{get forecast probability bins from set of forecasts probabilities}
\usage{
get_fc_prob_bins(event_tbl)
}
\description{
get forecast probability bins from set of forecasts probabilities
}
