% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_ensemble_verification.R
\name{get_marginal_fc_dist}
\alias{get_marginal_fc_dist}
\title{marginal distribution of the forecasts as in wilks page 285 eq 7.36 and 7.37}
\usage{
get_marginal_fc_dist(xfc_bins, event_tbl)
}
\description{
marginal distribution of the forecasts as in wilks page 285 eq 7.36 and 7.37
}
