% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_ensemble_verification.R
\name{getHRi_rel}
\alias{getHRi_rel}
\title{notation and expression comes from http://www.wmo.int/pages/prog/www/DPS/LRF/ATTACHII-8SVSfrom%20WMO_485_Vol_I.pdf}
\usage{
getHRi_rel(xfc_bins, Oi, NOi)
}
\description{
notation and expression comes from http://www.wmo.int/pages/prog/www/DPS/LRF/ATTACHII-8SVSfrom%20WMO_485_Vol_I.pdf
}
