% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_if_missing.R
\name{install_if_missing}
\alias{install_if_missing}
\title{install package if missing}
\usage{
install_if_missing(p)
}
\arguments{
\item{pkg}{the package name as string}
}
\description{
This function allows you to install a package if it hasn't been installed yet.
}
\examples{
install_if_missing("roxygen2")
}
\keyword{install}
\keyword{package}
