% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_hydraulics.R
\name{Q_out_muskingum}
\alias{Q_out_muskingum}
\title{Muskingum's effluent in pipe or catchment reach as in the documentation of citydrain 2}
\usage{
Q_out_muskingum(Qin, V, parm.df, i)
}
\arguments{
\item{Qin}{is a SpatialPointsDataFrame inflows to reach in step i}

\item{V}{is a SpatialPointsDataFrame of volumes in reach in step i-1}

\item{parm.df}{is SpatialPointsDataFrame pointing to the centroids of the catchment and containing dt, K, X and the id of the upstream subbasin}

\item{i}{is the time step}

\item{Qout}{is a SpatialPointsDataFrame of outflows from reach in step i}
}
\description{
Muskingum's effluent in pipe or catchment reach as in the documentation of citydrain 2
}
