% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_hydraulics.R
\name{Qoverflow_ret_str}
\alias{Qoverflow_ret_str}
\title{Overflow from retention structure}
\usage{
Qoverflow_ret_str(Qin, Qout, Vprevious, Vvirtual, parm.df)
}
\arguments{
\item{Qin}{is a SpatialPointsDataFrame inflow to structure in step i}

\item{Qout}{is a SpatialPointsDataFrame regulated outflow from structure in step i}

\item{Vprevious}{is the volume in step i-1}

\item{Vvirtual}{is the virtual volume in step i, including overflow volume}

\item{param.df}{is a SpatialPointsDataFrame of parameters of each subbasin, including df (time step), Vmax (maximum allowable volume of structure) and Qoutmax (maximum regulated allowable outflow from structure)}
}
\description{
Overflow from retention structure
}
