% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_ensemble_verification.R
\name{roc}
\alias{roc}
\title{roc of ensemble forecast}
\usage{
roc(xfc, xobs, type = "cont", prctile_upper = 0.33, prctile_lower = NULL,
  thresh_upper = NULL, thresh_lower = NULL)
}
\arguments{
\item{xfc}{is a dataframe with the forecast, each column corresponds to one ensemble member}

\item{xobs}{is a dataframe with the observations, the row order must be the same as in xfc}

\item{prctile_upper}{is the value of the upper percentile for defining an event}

\item{prctile_lower}{is the value of the lower percentile defining an event}

\item{thresh_upper}{is optional. In case no percentiles are given, a threshold value must be given}

\item{thresh_lower}{is optional. In case no percentiles are given, a threshold value must be given}
}
\description{
roc of ensemble forecast
}
