% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_WASA.R
\name{spacetime2WASA_P}
\alias{spacetime2WASA_P}
\title{writes WASA input file}
\usage{
spacetime2WASA_P(stObj, address)
}
\arguments{
\item{stObj}{is a spacetime object whose points are the centroids of the subbasins}

\item{address}{is the path where the file should be written}
}
\description{
writes WASA input file
}
