% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_hydraulics.R
\name{V_muskingum}
\alias{V_muskingum}
\title{Stored volume in step i. from the documentation of citydrain 2}
\usage{
V_muskingum(Qin, Qe, V, parm.df)
}
\arguments{
\item{Qin}{is a SpatialPointsDataFrame inflows to reach in step i}

\item{V}{is a SpatialPointsDataFrame of volumes in reach in step i-1}

\item{parm.df}{is SpatialPointsDataFrame pointing to the centroids of the catchment and containing dt, K, X and the id of the upstream subbasin}

\item{Qout}{is a SpatialPointsDataFrame of outflows from reach in step i}

\item{i}{is the time step}
}
\description{
Stored volume in step i. from the documentation of citydrain 2
}
