% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_hydraulics.R
\name{Virtual_retention}
\alias{Virtual_retention}
\title{Virtual volume in retention structure}
\usage{
Virtual_retention(Qin, Qoverflow, Vprevious, parm.df)
}
\arguments{
\item{Qin}{is a SpatialPointsDataFrame inflow to structure in step i}

\item{Qoverflow}{is a SpatialPointsDataFrame unregulated overflow from structure in step i}

\item{Vprevious}{is the volume in step i-1}

\item{Vvirtual}{is the virtual volume in step i, including overflow volume}

\item{param.df}{is a SpatialPointsDataFrame of parameters of each subbasin, including df (time step), Vmax (maximum allowable volume of structure) and Qoutmax (maximum regulated allowable outflow from structure)}
}
\description{
Virtual volume in retention structure
}
